backend/requirements.txt
fastapi[all]
uvicorn
python-jose[cryptography]
bcrypt==4.0.1
passlib
python-multipart
SQLAlchemy